+++ services/s3/e2eTest/src/commonMain/kotlin/aws/sdk/kotlin/services/s3/S3ExpressTest.kt
@@ -45,8 +45,13 @@ class S3ExpressTest {
 
     @AfterAll
     fun cleanup(): Unit = runBlocking {
-        testBuckets.forEach { bucket -> S3TestUtils.deleteBucket(client, bucket) }
-        client.close()
+        try {
+            if (::testBuckets.isInitialized) {
+                testBuckets.forEach { bucket -> S3TestUtils.deleteBucket(client, bucket) }
+            }
+        } finally {
+            client.close()
+        }
     }
 
     @Test
